Summarize scLinaX results (return per-individual summary)

Summarize scLinaX results (return per-individual summary)

Usage

summarize_scLinaX(scLinaX_obj, QC_total_allele_THR = 10, Annotation = NULL)

Arguments

scLinaX_obj

A list object generated by the run_scLinaX function.

QC_total_allele_THR

Threshold for the total allele count (depth) of the SNP used for calculating the ratio of expression from Xi. Note that this count is calculated with cells successfully assigned to the group based on the inactivated X chromosome. This filter is applied in the final step of scLinax and differs from "SNP_DETECTION_DP". Default: 10.

Annotation

A dataframe (tibble) for the annotation of the cells. This dataframe should have the following columns:

cell_barcode

Annotation

Value

A dataframe (tibble) containing scLinaX results.

If no Annotation is supplied, this function returns the ratio of expression from the inactivated X chromosome for all cells. If Annotation is supplied, this function returns the ratio of expression from the inactivated X chromosome for each cell annotation.

The dataframe includes the following columns:

Sample_ID: Sample ID

SNP_ID: SNP identifier

CHR: PAR or nonPAR

POS: Genomic position of the SNP (GRCh38)

REF: Reference allele of the SNP A,T,G,C

ALT: Alternative allele of the SNP A,T,G,C

Gene: Gene name

XCI_status: XCI status escape, variable, inactive, unknown

Gene_class: XCI status combined with CHR information PAR1, nonPAR_escape, nonPAR_variable, nonPAR_inactive, nonPAR_unknown, PAR2

Used_as_refGene: Whether the Gene was used as a reference gene Yes, No

Used_as_refSNP: Whether the SNP_ID was used as a reference SNP Yes, No

Total_A_allele, Total_B_allele: Total allele count of allele A and B

Total_allele: Total allele count of the SNP

Expressing_cells: Number of cells expressing the SNP

minor_allele_ratio: Ratio of the expression from the allele (A, B) with lower expression (=Ratio of the expression from Xi)

Reference_Gene, Reference_SNP: A list of genes used as reference genes and SNPs

Num_Reference_Gene, Num_Reference_SNP: Number of reference genes and SNPs

Reference_Cell_Count: Number of cells showing mono-allelic expression of reference SNPs

Num_A_cells, Num_B_cells: Number of cells showing mono-allelic expression of A and B alleles of reference SNPs

Num_Fail_cells: Number of cells showing bi-allelic expression of reference SNPs (should be removed from the analysis)

If in per-cell annotation mode, it includes these additional columns:

Major_allele: The allele with the higher read count A, B

Annotation: Cell annotation