Run scLinaX-multi using the scATAC data and output from scLinaX (Currently, only single sample mode is implemented).

Run scLinaX-multi using the scATAC data and output from scLinaX (Currently, only single sample mode is implemented).

Usage

run_scLinaX_multi(
  scLinaX_summary,
  scLinaX_obj,
  ATAC_ASE_data,
  peak_data,
  SNP_DETECTION_DP = 30,
  SNP_DETECTION_MAF = 0.1
)

Arguments

scLinaX_summary

Output from the summarize_scLinaX function run with Annotaion=NULL option

scLinaX_obj

A list object generated by the run_scLinaX function.

ATAC_ASE_data

A dataframe (tibble) containing single-cell chromatin accessibility data. This dataframe should include the following columns:

SNP_ID: SNP identifier

POS: Genomic position of the SNP (GRCh38)

REF: Reference allele of the SNP A,T,G,C

ALT: Alternative allele of the SNP A,T,G,C

cell_barcode: Cell barcode

REFcount: Allelic read count of the reference allele

ALTcount: Allelic read count of the alternative allele

OTHcount: Allelic read count of the other allele

Sample_ID: Sample ID

peak_data

A dataframe (tibble) containing peak information called from the single-cell chromatin accessibility data. This dataframe should include the following columns:

start: Genomic position of the start of the peak (1-base)

end: Genomic position of the end of the peak (1-base)

Peak_name: Name of the peak

Gene: Nearest gene to the peak

SNP_DETECTION_DP

Threshold for the total allele count (depth) of the SNP in the scATAC data. SNP–Sample pairs with a total allele count of at least "SNP_DETECTION_DP" are used for the analysis. Default: 30.

SNP_DETECTION_MAF

Threshold for the minor allele count of the scASE data. SNP–Sample pairs with a minor allele ratio between "SNP_DETECTION_MAF" and "1 - SNP_DETECTION_MAF" are used for the analysis. Default: 0.1.

Value

A dataframe (tibble) containing raw per-cell scLinaX-multi results.

This dataframe should include the following columns:

cell_barcode: Cell barcode

Sample_ID: Sample ID

SNP_ID: SNP identifier

CHR: PAR or nonPAR

POS: Genomic position of the SNP (GRCh38)

REF: Reference allele of the SNP A,T,G,C

ALT: Alternative allele of the SNP A,T,G,C

REFcount: Allelic read count of the reference allele

ALTcount: Allelic read count of the alternative allele

OTHcount: Allelic read count of the other allele

Peak_name: Name of the peak

Gene: Nearest gene to the peak

XCI_status: XCI status escape, variable, inactive, unknown

Gene_class: XCI status combined with CHR information PAR1, nonPAR_escape, nonPAR_variable, nonPAR_inactive, nonPAR_unknown, PAR2

Xa: Information of the activated X chromosome Allele_A, Allele_B

A_allele_count, B_allele_count: Total allele count of allele A and B